Refactor and expand test coverage
Refactored test files to improve clarity, modularity, and maintainability. Added new tests for `Clock`, `TaskManager`, `LiteSerializer`, and `Nalix.Network` components. Removed redundant tests and consolidated shared logic. Updated `Directory.Build.props` to centralize test dependencies. Enhanced README with test structure and guidelines.
