+++ tests/Nalix.Framework.Tests/Cryptography/ChaCha20Poly1305Tests.cs
@@ -1,5 +1,6 @@
 // Copyright (c) 2025-2026 PPN Corporation. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
+
 using Nalix.Framework.Security.Aead;
 using Xunit;
 
@@ -8,7 +9,7 @@ namespace Nalix.Framework.Tests.Cryptography;
 /// <summary>
 /// Unit tests for ChaCha20Poly1305 AEAD implementation.
 /// </summary>
-public class ChaCha20Poly1305Tests
+public sealed class ChaCha20Poly1305Tests
 {
     private static byte[] RandomBytes(int length)
     {
@@ -21,9 +22,8 @@ public class ChaCha20Poly1305Tests
     /// Verifies round-trip correctness using the Span-based API.
     /// </summary>
     [Fact]
-    public void EncryptDecryptRoundTripSpan()
+    public void EncryptThenDecryptWithSpanApiRoundTripsPayload()
     {
-        // Arrange
         byte[] key = RandomBytes(32);
         byte[] nonce = RandomBytes(12);
         byte[] plaintext = RandomBytes(128);
@@ -33,15 +33,12 @@ public class ChaCha20Poly1305Tests
         byte[] tag = new byte[ChaCha20Poly1305.TagSize];
         byte[] recovered = new byte[plaintext.Length];
 
-        // Act: encrypt (Span API)
         int ctWritten = ChaCha20Poly1305.Encrypt(key, nonce, plaintext, aad, ciphertext, tag);
         Assert.Equal(plaintext.Length, ctWritten);
 
-        // Act: decrypt (Span API)
         int ptWritten = ChaCha20Poly1305.Decrypt(key, nonce, ciphertext, aad, tag, recovered);
         Assert.Equal(plaintext.Length, ptWritten);
 
-        // Assert: plaintext matches
         Assert.Equal(plaintext, recovered);
     }
 
@@ -49,9 +46,8 @@ public class ChaCha20Poly1305Tests
     /// Tampering with the authentication tag should cause Span-based decryption to return a negative value.
     /// </summary>
     [Fact]
-    public void DecryptTamperedTagReturnsNegativeSpan()
+    public void DecryptWhenTagIsTamperedReturnsNegativeResult()
     {
-        // Arrange
         byte[] key = RandomBytes(32);
         byte[] nonce = RandomBytes(12);
         byte[] plaintext = RandomBytes(32);
@@ -61,17 +57,11 @@ public class ChaCha20Poly1305Tests
         byte[] tag = new byte[ChaCha20Poly1305.TagSize];
 
         _ = ChaCha20Poly1305.Encrypt(key, nonce, plaintext, aad, ciphertext, tag);
-
-        // Tamper with tag
         tag[0] ^= 0xFF;
-
-        // Destination buffer for plaintext
         byte[] recovered = new byte[plaintext.Length];
 
-        // Act
         int result = ChaCha20Poly1305.Decrypt(key, nonce, ciphertext, aad, tag, recovered);
 
-        // Assert: decryption should fail (method returns negative according to implementation)
         Assert.True(result < 0, "Span-based Decrypt should return a negative value on authentication failure.");
     }
 }
