+++ tests/Nalix.Framework.Tests/Serialization/LiteSerializerUnmanagedTests.cs
@@ -7,14 +7,12 @@ using Xunit;
 
 namespace Nalix.Framework.Tests.Serialization;
 
-public class LiteSerializerUnmanagedTests
+public sealed class LiteSerializerUnmanagedTests
 {
     [Fact]
-    public void SerializeDeserializeUnmanagedStructRoundTrip()
+    public void SerializeDeserialize_UnmanagedStruct_RoundTripsValue()
     {
         ComplexStruct input = new() { I32 = 123456789, I16 = -1234, B = 0xAB };
-
-        // byte[] path (unmanaged -> không cần formatter)
         byte[] data = LiteSerializer.Serialize(in input);
 
         ComplexStruct output = default;
@@ -27,46 +25,42 @@ public class LiteSerializerUnmanagedTests
     }
 
     [Fact]
-    public void SerializeDeserializeNullClassRoundTrip()
+    public void SerializeDeserialize_NullClass_RoundTripsValue()
     {
-        NullClass input = new() { I32 = null, I16 = null };
-
-        // byte[] path (unmanaged -> không cần formatter)
+        NullClass input = new() { I32 = [], I16 = null };
         byte[] data = LiteSerializer.Serialize(in input);
 
-        NullClass output = default;
+        NullClass? output = null;
         int read = LiteSerializer.Deserialize(data, ref output);
 
         Assert.Equal(data.Length, read);
+        Assert.NotNull(output);
         Assert.Equal(input.I32, output.I32);
         Assert.Equal(input.I16, output.I16);
     }
 
     [Fact]
-    public void SerializeUnmanagedStructToProvidedBufferExactSizeOk()
+    public void SerializeToProvidedBuffer_UnmanagedStruct_ExactBuffer_WritesSuccessfully()
     {
         ComplexStruct value = new() { I32 = 42, I16 = 7, B = 1 };
-        // Lấy size từ serialize để có kích cỡ tối thiểu
-        byte[] tmp = LiteSerializer.Serialize(in value);
-        byte[] buffer = new byte[tmp.Length];
+        byte[] exact = LiteSerializer.Serialize(in value);
+        byte[] buffer = new byte[exact.Length];
 
         int written = LiteSerializer.Serialize(in value, buffer);
-        Assert.Equal(tmp.Length, written);
+        ComplexStruct output = default;
+        int read = LiteSerializer.Deserialize(buffer, ref output);
 
-        ComplexStruct back = default;
-        int read = LiteSerializer.Deserialize(buffer, ref back);
-
+        Assert.Equal(exact.Length, written);
         Assert.Equal(buffer.Length, read);
-        Assert.Equal(value.I32, back.I32);
-        Assert.Equal(value.I16, back.I16);
-        Assert.Equal(value.B, back.B);
+        Assert.Equal(value.I32, output.I32);
+        Assert.Equal(value.I16, output.I16);
+        Assert.Equal(value.B, output.B);
     }
 
     [Fact]
-    public void SerializeUnmanagedStructBufferTooSmallThrows()
+    public void SerializeToProvidedBuffer_UnmanagedStruct_ShortBuffer_ThrowsSerializationFailureException()
     {
         ComplexStruct value = new() { I32 = 1, I16 = 2, B = 3 };
-        // Lấy mảng đúng size rồi cắt nhỏ đi 1 byte
         byte[] exact = LiteSerializer.Serialize(in value);
         byte[] tooSmall = new byte[exact.Length - 1];
 
@@ -74,16 +68,14 @@ public class LiteSerializerUnmanagedTests
     }
 
     [Fact]
-    public void DeserializeUnmanagedStructBufferTooSmallThrows()
+    public void Deserialize_UnmanagedStruct_ShortBuffer_ThrowsSerializationFailureException()
     {
         ComplexStruct value = new() { I32 = 1, I16 = 2, B = 3 };
         byte[] full = LiteSerializer.Serialize(in value);
-        byte[] small = new byte[full.Length - 1];
+        byte[] shortBuffer = new byte[full.Length - 1];
+        Array.Copy(full, shortBuffer, shortBuffer.Length);
 
-        // copy thiếu dữ liệu
-        Array.Copy(full, small, small.Length);
-
-        ComplexStruct dest = default;
-        _ = Assert.Throws<SerializationFailureException>(() => LiteSerializer.Deserialize(small, ref dest));
+        ComplexStruct destination = default;
+        _ = Assert.Throws<SerializationFailureException>(() => LiteSerializer.Deserialize(shortBuffer, ref destination));
     }
 }
