+++ tests/Nalix.Framework.Tests/Time/ClockValidationTests.cs
@@ -1,6 +1,8 @@
 // Copyright (c) 2025-2026 PPN Corporation. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
+
 using System;
+using System.Threading;
 using Nalix.Framework.Time;
 using Xunit;
 
@@ -10,10 +12,10 @@ namespace Nalix.Framework.Tests.Time;
 /// Tests for input validation in Clock operations.
 /// </summary>
 [Collection("ClockTests")]
-public class ClockValidationTests
+public sealed class ClockValidationTests
 {
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectNegativeServerTime()
+    public void SynchronizeUnixMillisecondsRejectsNegativeServerTime()
     {
         Clock.ResetSynchronization();
 
@@ -25,7 +27,7 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectNegativeRTT()
+    public void SynchronizeUnixMillisecondsRejectsNegativeRtt()
     {
         Clock.ResetSynchronization();
 
@@ -39,7 +41,7 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectZeroMaxAllowedDrift()
+    public void SynchronizeUnixMillisecondsRejectsZeroMaxAllowedDrift()
     {
         Clock.ResetSynchronization();
         long currentUnixMs = (long)(DateTime.UtcNow - DateTime.UnixEpoch).TotalMilliseconds;
@@ -52,7 +54,7 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectNegativeMaxAllowedDrift()
+    public void SynchronizeUnixMillisecondsRejectsNegativeMaxAllowedDrift()
     {
         Clock.ResetSynchronization();
         long currentUnixMs = (long)(DateTime.UtcNow - DateTime.UnixEpoch).TotalMilliseconds;
@@ -65,7 +67,7 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectZeroMaxHardAdjust()
+    public void SynchronizeUnixMillisecondsRejectsZeroMaxHardAdjust()
     {
         Clock.ResetSynchronization();
         long currentUnixMs = (long)(DateTime.UtcNow - DateTime.UnixEpoch).TotalMilliseconds;
@@ -78,7 +80,7 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectNegativeMaxHardAdjust()
+    public void SynchronizeUnixMillisecondsRejectsNegativeMaxHardAdjust()
     {
         Clock.ResetSynchronization();
         long currentUnixMs = (long)(DateTime.UtcNow - DateTime.UnixEpoch).TotalMilliseconds;
@@ -91,11 +93,10 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectTimestampBeforeYear2000()
+    public void SynchronizeUnixMillisecondsRejectsTimestampBeforeYear2000()
     {
         Clock.ResetSynchronization();
 
-        // Unix timestamp for Jan 1, 1999 00:00:00
         const long oldTimestamp = 915148800000L;
 
         ArgumentException ex = Assert.Throws<ArgumentException>(() =>
@@ -106,11 +107,10 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldRejectTimestampAfterYear2100()
+    public void SynchronizeUnixMillisecondsRejectsTimestampAfterYear2100()
     {
         Clock.ResetSynchronization();
 
-        // Unix timestamp for Jan 1, 2101 00:00:00
         const long futureTimestamp = 4133980800000L;
 
         ArgumentException ex = Assert.Throws<ArgumentException>(() =>
@@ -121,66 +121,46 @@ public class ClockValidationTests
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldAcceptValidCurrentTimestamp()
+    public void SynchronizeUnixMillisecondsAcceptsReasonableCurrentTimestamp()
     {
         Clock.ResetSynchronization();
-
-        // Wait a bit to ensure clean state
-        System.Threading.Thread.Sleep(10);
-
-        // Use a time slightly in the future to ensure drift threshold is exceeded
+        Thread.Sleep(10);
         long futureUnixMs = (long)(DateTime.UtcNow.AddSeconds(10) - DateTime.UnixEpoch).TotalMilliseconds;
-
-        // Should not throw - use thresholds that will trigger sync
         double adjustment = Clock.SynchronizeUnixMilliseconds(futureUnixMs, 10, 1000, 50000);
 
         Assert.True(Clock.IsSynchronized);
-        // The adjustment should be around 10 seconds (10000 ms)
         Assert.InRange(Math.Abs(adjustment), 9000, 11000);
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldHandleEdgeCaseYear2000()
+    public void SynchronizeUnixMillisecondsAcceptsYear2000Boundary()
     {
         Clock.ResetSynchronization();
 
-        // Unix timestamp for Jan 1, 2000 00:00:00 (exactly at boundary)
         const long y2kTimestamp = 946684800000L;
-
-        // Should not throw
         _ = Clock.SynchronizeUnixMilliseconds(y2kTimestamp, 0, 100000, 100000);
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsShouldHandleEdgeCaseYear2100()
+    public void SynchronizeUnixMillisecondsAcceptsYear2100Boundary()
     {
         Clock.ResetSynchronization();
 
-        // Unix timestamp for Jan 1, 2100 00:00:00 (exactly at boundary)
         const long y2100Timestamp = 4102444800000L;
-
-        // Should not throw
         _ = Clock.SynchronizeUnixMilliseconds(y2100Timestamp, 0, 100000, 100000);
     }
 
     [Fact]
-    public void SynchronizeUnixMillisecondsWithRTTShouldAdjustTimeCorrectly()
+    public void SynchronizeUnixMillisecondsAccountsForRoundTripTime()
     {
         Clock.ResetSynchronization();
-
-        // Wait a bit to ensure clean state
-        System.Threading.Thread.Sleep(10);
-
-        // Use a time slightly in the future to ensure drift threshold is exceeded
+        Thread.Sleep(10);
         long futureUnixMs = (long)(DateTime.UtcNow.AddSeconds(10) - DateTime.UnixEpoch).TotalMilliseconds;
-        const double rttMs = 100; // 100ms round trip
+        const double rttMs = 100;
 
-        // The method should add half of RTT (50ms) to the server time
-        // Use thresholds that will trigger sync
         double adjustment = Clock.SynchronizeUnixMilliseconds(futureUnixMs, rttMs, 1000, 50000);
 
         Assert.True(Clock.IsSynchronized);
-        // The adjustment should be around 10 seconds + half RTT (10050 ms)
         Assert.InRange(Math.Abs(adjustment), 9000, 11000);
     }
 }
