+++ tests/Nalix.Network.Test/Nalix.Network.Tests.csproj
@@ -6,12 +6,6 @@
 
 	<ItemGroup>
 		<PackageReference Include="FluentAssertions" Version="8.9.0" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.3.0" />
-		<PackageReference Include="xunit" Version="2.9.3" />
-		<PackageReference Include="xunit.runner.visualstudio" Version="3.1.5">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
 		<PackageReference Include="Xunit.SkippableFact" Version="1.5.61" />
 	</ItemGroup>
 
@@ -19,6 +13,8 @@
 		<ProjectReference Include="..\..\src\Nalix.Common\Nalix.Common.csproj" />
 		<ProjectReference Include="..\..\src\Nalix.Framework\Nalix.Framework.csproj" />
 		<ProjectReference Include="..\..\src\Nalix.Logging\Nalix.Logging.csproj" />
+		<ProjectReference Include="..\..\src\Nalix.Network\Nalix.Network.csproj" />
+		<ProjectReference Include="..\..\src\Nalix.Network.Pipeline\Nalix.Network.Pipeline.csproj" />
 	</ItemGroup>
 
 </Project>
